ai-server/requirements.txt
Flask==3.0.0
openai==1.51.2
pymongo==4.10.1
pydantic==2.9.2
python-dotenv==1.0.1